gestion conflit

+++ app/src/main/java/fr/dut/ptut2021/models/Point.java
@@ -2,14 +2,14 @@ package fr.dut.ptut2021.models;
 
 public class Point {
 
-    private int x,y;
+    private double x,y;
 
     public Point(){
         this.x = 0;
         this.y = 0;
     }
 
-    public Point(int x, int y){
+    public Point(double x, double y){
         this.x = x;
         this.y = y;
     }
@@ -19,19 +19,19 @@ public class Point {
         this.y = p.y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 }
